Make bdk_file_store use bincode v1
