+++ crates/file_store/Cargo.toml
@@ -7,7 +7,7 @@ repository = "https://github.com/bitcoindevkit/bdk"
 
 [dependencies]
 bdk_chain = { path = "../chain", version = "0.3.1", features = [ "serde", "miniscript" ] }
-bincode = { version = "2.0.0-rc.2", features = [ "serde" ] }
+bincode = { version = "1" }
 serde = { version = "1", features = ["derive"] }
 
 [dev-dependencies]
